model defined
